Introduce zooming custom attribute
